Add more precision to float display

+++ WowPacketParser/Misc/PacketReads.cs
@@ -728,7 +728,7 @@ namespace WowPacketParser.Misc
         private static string FormatFloat(float value)
         {
             if (!Settings.DebugReads)
-                return string.Format("{0:F20}", value).Substring(0, 20).TrimEnd('0').TrimEnd('.');
+                return string.Format("{0:F20}", (double) value).Substring(0, 20).TrimEnd('0').TrimEnd('.');
 
             var bytes = BitConverter.GetBytes(value);
             return value + " (0x" + BitConverter.ToString(bytes) + ")";
